refactored the whole skos functionaliuty of v1 and v2 for EBISPOT#625 and TIBHAnnover#1
