+++ backend/src/main/java/uk/ac/ebi/spot/ols/controller/api/v2/V2OntologySKOSConceptController.java
@@ -8,8 +8,6 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
@@ -25,22 +23,21 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.util.UriUtils;
 import uk.ac.ebi.spot.ols.controller.api.v1.TopConceptEnum;
-import uk.ac.ebi.spot.ols.controller.api.v1.V1TermAssembler;
-import uk.ac.ebi.spot.ols.model.v1.V1Term;
-import uk.ac.ebi.spot.ols.model.v2.SKOSRelation;
+import uk.ac.ebi.spot.ols.model.Edge;
+import uk.ac.ebi.spot.ols.model.Node;
+import uk.ac.ebi.spot.ols.model.SKOSRelation;
 import uk.ac.ebi.spot.ols.model.v2.V2Entity;
 import uk.ac.ebi.spot.ols.repository.v1.TreeNode;
-import uk.ac.ebi.spot.ols.repository.v1.V1TermRepository;
-import uk.ac.ebi.spot.ols.repository.v2.V2ClassRepository;
+import uk.ac.ebi.spot.ols.repository.v2.V2SKOSRepository;
 
 import javax.servlet.http.HttpServletRequest;
 import java.io.IOException;
 import java.util.*;
 
 /**
- * @author Simon Jupp
- * @date 02/11/15
- * Samples, Phenotypes and Ontologies Team, EMBL-EBI
+ * @author Erhun Giray TUNCAY
+ * @email giray.tuncay@tib.eu
+ * TIB-Leibniz Information Center for Science and Technology
  */
 @RestController
 @RequestMapping("/api/v2/ontologies")
@@ -48,7 +45,7 @@ import java.util.*;
 public class V2OntologySKOSConceptController {
 
     @Autowired
-    V2ClassRepository classRepository;
+    V2SKOSRepository skosRepository;
 
     @Operation(description = "Get complete SKOS concept hierarchy or only top concepts based on alternative top concept identification methods and concept relations. If only top concepts are identified, they can be used to extract the following levels of the concept tree one by one using the /{onto}/conceptrelations/{iri} method with broader or narrower concept relations.")
     @RequestMapping(path = "/{onto}/skos/tree", produces = {MediaType.APPLICATION_JSON_VALUE, MediaTypes.HAL_JSON_VALUE}, method = RequestMethod.GET)
@@ -66,9 +63,9 @@ public class V2OntologySKOSConceptController {
             Pageable pageable) throws IOException {
     	ontologyId = ontologyId.toLowerCase();
     	if (TopConceptEnum.RELATIONSHIPS == topConceptIdentification)
-    		return new ResponseEntity<>(classRepository.conceptTreeWithoutTop(ontologyId,narrower,withChildren,obsoletes,lang,pageable), HttpStatus.OK);
+    		return new ResponseEntity<>(skosRepository.conceptTreeWithoutTop(ontologyId,narrower,withChildren,obsoletes,lang,pageable), HttpStatus.OK);
     	else
-    		return new ResponseEntity<>(classRepository.conceptTree(ontologyId,TopConceptEnum.SCHEMA == topConceptIdentification,narrower, withChildren,obsoletes,lang,pageable), HttpStatus.OK);
+    		return new ResponseEntity<>(skosRepository.conceptTree(ontologyId,TopConceptEnum.SCHEMA == topConceptIdentification,narrower, withChildren,obsoletes,lang,pageable), HttpStatus.OK);
     }
 
     @Operation(description = "Display complete SKOS concept hierarchy or only top concepts based on alternative top concept identification methods and concept relations. If only top concepts are identified, they can be used to extract the following levels of the concept tree one by one using the /{onto}/displayconceptrelations/{iri} method with broader or narrower concept relations.")
@@ -91,9 +88,9 @@ public class V2OntologySKOSConceptController {
     	 ontologyId = ontologyId.toLowerCase();
      	 List<TreeNode<V2Entity>> rootIndividuals = null;
     	 if(TopConceptEnum.RELATIONSHIPS == topConceptIdentification)
-    		 rootIndividuals = classRepository.conceptTreeWithoutTop(ontologyId,narrower,withChildren,obsoletes,lang,pageable);
+    		 rootIndividuals = skosRepository.conceptTreeWithoutTop(ontologyId,narrower,withChildren,obsoletes,lang,pageable);
     	 else
-    		 rootIndividuals = classRepository.conceptTree(ontologyId,TopConceptEnum.SCHEMA == topConceptIdentification,narrower, withChildren,obsoletes,lang,pageable);
+    		 rootIndividuals = skosRepository.conceptTree(ontologyId,TopConceptEnum.SCHEMA == topConceptIdentification,narrower, withChildren,obsoletes,lang,pageable);
          StringBuilder sb = new StringBuilder();
          for (TreeNode<V2Entity> root : rootIndividuals) {
         	 sb.append(root.getIndex() + " , "+ root.getData().any().get("label").toString() + " , " + root.getData().any().get("iri").toString()).append("\n");
@@ -121,7 +118,7 @@ public class V2OntologySKOSConceptController {
     	TreeNode<V2Entity> topConcept = new TreeNode<V2Entity>(new V2Entity(new JsonObject()));
     	String decodedIri;
 		decodedIri = UriUtils.decode(iri, "UTF-8");
-		topConcept = classRepository.conceptSubTree(ontologyId, decodedIri, narrower, index, obsoletes, lang, pageable);
+		topConcept = skosRepository.conceptSubTree(ontologyId, decodedIri, narrower, index, obsoletes, lang, pageable);
 
         if (topConcept.getData().any().get("iri").toString() == null)
             throw new ResourceNotFoundException("No roots could be found for " + ontologyId );
@@ -150,7 +147,7 @@ public class V2OntologySKOSConceptController {
 	    	String decodedIri;
 	    	StringBuilder sb = new StringBuilder();
 			decodedIri = UriUtils.decode(iri, "UTF-8");
-			topConcept = classRepository.conceptSubTree(ontologyId, decodedIri, narrower, index, obsoletes, lang, pageable);
+			topConcept = skosRepository.conceptSubTree(ontologyId, decodedIri, narrower, index, obsoletes, lang, pageable);
 
         	sb.append(topConcept.getIndex() + " , "+ topConcept.getData().any().get("label").toString() + " , " + topConcept.getData().any().get("iri").toString()).append("\n");
 	        sb.append(generateConceptHierarchyTextByOntology(topConcept, displayRelated));
@@ -176,7 +173,7 @@ public class V2OntologySKOSConceptController {
     	ontologyId = ontologyId.toLowerCase();
     	List<V2Entity> related = new ArrayList<V2Entity>();
 		String decodedIri = UriUtils.decode(iri, "UTF-8");
-		related = classRepository.findRelated(ontologyId, decodedIri, relationType.getPropertyName(),lang);
+		related = skosRepository.findRelated(ontologyId, decodedIri, relationType.getPropertyName(),lang);
 
         final int start = (int)pageable.getOffset();
         final int end = Math.min((start + pageable.getPageSize()), related.size());
@@ -205,7 +202,7 @@ public class V2OntologySKOSConceptController {
     	ontologyId = ontologyId.toLowerCase();
     	List<V2Entity> related = new ArrayList<V2Entity>();
 		String decodedIri = UriUtils.decode(iri, "UTF-8");
-		related = classRepository.findRelated(ontologyId, decodedIri, relationType.getPropertyName(),lang);
+		related = skosRepository.findRelated(ontologyId, decodedIri, relationType.getPropertyName(),lang);
 
         final int start = (int)pageable.getOffset();
         final int end = Math.min((start + pageable.getPageSize()), related.size());
@@ -234,7 +231,7 @@ public class V2OntologySKOSConceptController {
     	ontologyId = ontologyId.toLowerCase();
     	List<V2Entity> related = new ArrayList<V2Entity>();
 		String decodedIri = UriUtils.decode(iri, "UTF-8");
-		related = classRepository.findRelatedIndirectly(ontologyId, decodedIri, relationType.getPropertyName(), obsoletes,lang,pageable);
+		related = skosRepository.findRelatedIndirectly(ontologyId, decodedIri, relationType.getPropertyName(), obsoletes,lang,pageable);
 
         return new ResponseEntity<>( related, HttpStatus.OK);
 
@@ -259,7 +256,7 @@ public class V2OntologySKOSConceptController {
     	ontologyId = ontologyId.toLowerCase();
     	List<V2Entity> related = new ArrayList<V2Entity>();
 		String decodedIri = UriUtils.decode(iri, "UTF-8");
-		related = classRepository.findRelatedIndirectly(ontologyId, decodedIri, relationType.getPropertyName(), obsoletes,lang,pageable);
+		related = skosRepository.findRelatedIndirectly(ontologyId, decodedIri, relationType.getPropertyName(), obsoletes,lang,pageable);
 
     	int count = 0;
         for (V2Entity individual : related)
@@ -282,15 +279,15 @@ public class V2OntologySKOSConceptController {
         List<V2Entity> related = new ArrayList<V2Entity>();
         String decodedIri = UriUtils.decode(iri, "UTF-8");
 
-        V2Entity subjectTerm = classRepository.findByOntologyAndIri(ontologyId, decodedIri, lang);
+        V2Entity subjectTerm = skosRepository.findByOntologyAndIri(ontologyId, decodedIri, lang);
 
-        related = classRepository.findRelated(ontologyId, decodedIri, SKOSRelation.related.getPropertyName(), lang);
+        related = skosRepository.findRelated(ontologyId, decodedIri, SKOSRelation.related.getPropertyName(), lang);
 
         List<V2Entity> narrower = new ArrayList<V2Entity>();
-        narrower = classRepository.findRelated(ontologyId, decodedIri, SKOSRelation.narrower.getPropertyName(), lang);
+        narrower = skosRepository.findRelated(ontologyId, decodedIri, SKOSRelation.narrower.getPropertyName(), lang);
 
         List<V2Entity> broader = new ArrayList<V2Entity>();
-        broader = classRepository.findRelated(ontologyId, decodedIri, SKOSRelation.broader.getPropertyName(), lang);
+        broader = skosRepository.findRelated(ontologyId, decodedIri, SKOSRelation.broader.getPropertyName(), lang);
 
         Set<Node> relatedNodes = new HashSet<Node>();
         related.forEach(term -> relatedNodes.add(new Node(term.any().get("iri").toString(), term.any().get("label").toString())));
@@ -300,9 +297,9 @@ public class V2OntologySKOSConceptController {
         broader.forEach(term -> broaderNodes.add(new Node(term.any().get("iri").toString(), term.any().get("label").toString())));
 
         Set<Edge> edges = new HashSet<Edge>();
-        relatedNodes.forEach(node -> edges.add(new Edge(decodedIri, node.iri, "related",SKOSRelation.related.getPropertyName())));
-        narrowerNodes.forEach(node -> edges.add(new Edge(decodedIri, node.iri, "narrower",SKOSRelation.narrower.getPropertyName())));
-        broaderNodes.forEach(node -> edges.add(new Edge(decodedIri, node.iri, "broader",SKOSRelation.broader.getPropertyName())));
+        relatedNodes.forEach(node -> edges.add(new Edge(decodedIri, node.getIri(), "related",SKOSRelation.related.getPropertyName())));
+        narrowerNodes.forEach(node -> edges.add(new Edge(decodedIri, node.getIri(), "narrower",SKOSRelation.narrower.getPropertyName())));
+        broaderNodes.forEach(node -> edges.add(new Edge(decodedIri, node.getIri(), "broader",SKOSRelation.broader.getPropertyName())));
 
         Set<Node> nodes = new HashSet<Node>();
         nodes.add(new Node(decodedIri,subjectTerm.any().get("label").toString()));
@@ -339,56 +336,6 @@ public class V2OntologySKOSConceptController {
     @ResponseStatus(value = HttpStatus.NOT_FOUND, reason = "Resource not found")
     @ExceptionHandler(ResourceNotFoundException.class)
     public void handleError(HttpServletRequest req, Exception exception) {
-    }
-
-    public class Node {
-        String iri;
-        String label;
-
-        public Node(String iri, String label) {
-            this.iri = iri;
-            this.label = label;
-        }
-
-        public String getIri() {
-            return iri;
-        }
-
-        public String getLabel() {
-            return label;
-        }
-
-    }
-
-    public class Edge {
-        String source;
-        String target;
-        String label;
-        String uri;
-
-        public Edge(String source, String target, String label, String uri) {
-            this.source = source;
-            this.target = target;
-            this.label = label;
-            this.uri = uri;
-        }
-
-        public String getSource() {
-            return source;
-        }
-
-        public String getTarget() {
-            return target;
-        }
-
-        public String getLabel() {
-            return label;
-        }
-
-        public String getUri() {
-            return uri;
-        }
-
     }
 
 }
